pyproject.toml
[project]
name = "sydeny_startup_agent"
version = "0.1.0"
description = "Agentic Webscrapper and RAG Analyst"
dependencies = [
    "crawl4ai>=0.7.8",
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.6",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "typing-extensions>=4.15.0",
]

[tool.setuptools]
packages = ["my_agent"]
